requirements.txt
opencv-python==4.8.1.78
customtkinter==5.2.0
Pillow==10.0.1
requests==2.31.0
psutil==5.9.6
pygame==2.5.2
numpy==1.24.3
mediapipe==0.10.7
pywin32==306
